python-client/requirements.txt
typer>=0.9.0
requests>=2.31.0
rich>=13.7.0
websocket-client>=1.7.0
python-dotenv>=1.0.0
